lp_website/backoffice/rest_api/serializers.py
from backoffice.models import SchoolClass
from rest_framework import serializers

class SchoolClassSerializer(serializers.ModelSerializer):
	class Meta:
		model = SchoolClass
		fields = ('id', 'name', 'school_name')