requirements.txt
Django==1.7
MySQL-python==1.2.5
django-bootstrap3==4.11.0
django_extensions==1.4.0
django-static-precompiler==0.8
django-awesome-avatar==1.1.3
pillow==2.7.0
django-imagekit==3.2.5
djangorestframework==3.0.5
